Get some of the async logger working.
Conceptually, I don't have a fantastic handle on the interplay between
the coroutines and the event loop. It seems like the event loop calls
the callbacks outside of the coroutine context that I wrapped things in
with the main function. My guess is that there is some C level
manipulation going here that I don't fully comprehend.